Move NodeLocationPreference to routing module
Move the NodeLocationPreference enum and its impl block from
policies::load_balancing::default to routing,
preparing for it to become a session-level configuration option.
The type is pub(crate) for now.
